no more need to fix an addres for udynlink_get_lot_base function
the asm template has been changed to load into the r1 register (to jump to udynlink_get_lot_base address) the value of the defined label.
then the user shall call the function udynlink_patch_exported_func to write the address of udynlink_get_lot_base to the space reserved for the label.
the only thing that shall be known at compile time is the code length in bytes of the asm prologue.
the rodata has been moved to the data section because the udynlink_patch_exported_func patch the global symbols belonging to the text section